requirements.txt
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
tk==0.1.0
